+++ src/routines.h
@@ -18,6 +18,7 @@ void C_player_add_exp(long num);
 
 signed char C_class_melee_bonus(enum class_t class);
 signed char C_class_ranged_bonus(enum class_t class);
+boolean C_class_can_use_item(enum class_t class, treasure_type const* item);
 
 uint8_t C_magic_spell_level(int32_t slot);
 uint8_t C_magic_spell_mana(int32_t slot);